+++ perl/LinkedSpec/Compiler.pm
@@ -49,87 +49,102 @@ sub _require_linkedre_pkg {
  return 1
 }
 
+sub _call_preserving_err {
+ my ($cb) = @_;
+ my $saved_err = $@;
+ my $wantarray = wantarray;
+ if ($wantarray) {
+  my @ret = $cb->();
+  $@ = $saved_err;
+  return @ret
+ }
+ if (defined $wantarray) {
+  my $ret = $cb->();
+  $@ = $saved_err;
+  return $ret
+ }
+ $cb->();
+ $@ = $saved_err;
+ return
+}
+
 sub _dump_value {
  my ($value) = @_;
- my $saved_err = $@;
- _require_data_dumper_pkg();
- my $ret = Data::Dumper::Dumper($value);
- $@ = $saved_err;
- return $ret
+ return _call_preserving_err(sub {
+  _require_data_dumper_pkg();
+  return Data::Dumper::Dumper($value)
+ })
 }
 
 sub _ored_re {
  my (@regexes) = @_;
- my $saved_err = $@;
- _require_linkedre_pkg();
- my $ret = LinkedRE::oredRE(@regexes);
- $@ = $saved_err;
- return $ret
+ return _call_preserving_err(sub {
+  _require_linkedre_pkg();
+  return LinkedRE::oredRE(@regexes)
+ })
 }
 
 sub _trace_log_output {
- my $saved_err = $@;
- _require_trace_pkg();
- my $ret = LinkedSpec::Trace::log_output(@_);
- $@ = $saved_err;
- return $ret
+ my @args = @_;
+ return _call_preserving_err(sub {
+  _require_trace_pkg();
+  return LinkedSpec::Trace::log_output(@args)
+ })
 }
 
 sub _trace_log_dump {
- my $saved_err = $@;
- _require_trace_pkg();
- my $ret = LinkedSpec::Trace::log_dump(@_);
- $@ = $saved_err;
- return $ret
+ my @args = @_;
+ return _call_preserving_err(sub {
+  _require_trace_pkg();
+  return LinkedSpec::Trace::log_dump(@args)
+ })
 }
 
 sub _trace_should_dump {
- my $saved_err = $@;
- return 0 unless exists $INC{'LinkedSpec/Trace.pm'};
- my $ret = LinkedSpec::Trace::should_dump(@_);
- $@ = $saved_err;
- return $ret
+ my @args = @_;
+ return _call_preserving_err(sub {
+  return 0 unless exists $INC{'LinkedSpec/Trace.pm'};
+  return LinkedSpec::Trace::should_dump(@args)
+ })
 }
 
 sub _trace_enter {
- my $saved_err = $@;
- _require_trace_pkg();
- my $ret = LinkedSpec::Trace::trace_enter(@_);
- $@ = $saved_err;
- return $ret
+ my @args = @_;
+ return _call_preserving_err(sub {
+  _require_trace_pkg();
+  return LinkedSpec::Trace::trace_enter(@args)
+ })
 }
 
 sub _trace_exit {
- my $saved_err = $@;
- _require_trace_pkg();
- my $ret = LinkedSpec::Trace::trace_exit(@_);
- $@ = $saved_err;
- return $ret
+ my @args = @_;
+ return _call_preserving_err(sub {
+  _require_trace_pkg();
+  return LinkedSpec::Trace::trace_exit(@args)
+ })
 }
 
 sub _trace_decision {
- my $saved_err = $@;
- _require_trace_pkg();
- my $ret = LinkedSpec::Trace::trace_decision(@_);
- $@ = $saved_err;
- return $ret
+ my @args = @_;
+ return _call_preserving_err(sub {
+  _require_trace_pkg();
+  return LinkedSpec::Trace::trace_decision(@args)
+ })
 }
 
 sub _trace_apply_trace_options {
  my ($option) = @_;
- my $saved_err = $@;
- _require_trace_pkg();
- my $ret = LinkedSpec::Trace::_apply_trace_options($option);
- $@ = $saved_err;
- return $ret
+ return _call_preserving_err(sub {
+  _require_trace_pkg();
+  return LinkedSpec::Trace::_apply_trace_options($option)
+ })
 }
 
 sub _trace_level_name_for_current_verbosity {
- my $saved_err = $@;
- _require_trace_pkg();
- my $ret = LinkedSpec::Trace::_trace_level_name($LinkedSpec::Trace::DUMP_VERBOSITY);
- $@ = $saved_err;
- return $ret
+ return _call_preserving_err(sub {
+  _require_trace_pkg();
+  return LinkedSpec::Trace::_trace_level_name($LinkedSpec::Trace::DUMP_VERBOSITY)
+ })
 }
 
 sub _default_bootstrap_parse_cb {